+++ HealthBar.h
@@ -10,7 +10,7 @@ namespace Game {
 	class HealthBar : public Entity{
 	public:
 
-		HealthBar(size_t startingHealth);
+		HealthBar(double startingHealth);
 
 		virtual SDL_Rect drawRect() const;
         virtual SDL_Rect collisionRect() const;
@@ -21,8 +21,8 @@ namespace Game {
 
         virtual Surface *image() const;
 
-		size_t getHealth() const;
-		void setHealth(int health);
+		double getHealth() const;
+		void setHealth(double health);
 
 		void update(double delta);
 		virtual void draw(Point offset = Point(), Surface &surface = screenBuf) const;
@@ -30,7 +30,7 @@ namespace Game {
 	private:
 		double fillPercent_;
 		double fillingPercent_;
-		size_t lastSetHealth_;
+		double lastSetHealth_;
 
 		pixels_t getMaxWidth() const;
 		pixels_t getFillWidth() const;
